refactor: :art: update import paths for user model and schema to improve module organization
